fixxing other merge conflicts

+++ src/main/java/gov/iti/jets/team5/repositories/ProductRepository.java
@@ -92,7 +92,7 @@ public class ProductRepository {
                     productDto.setProductID(String.valueOf(item.getId()));
                     productDto.setProductName(item.getProductName());
                     productDto.setProductImageURL("images/products/MeatlessPieces.png");
-                    productDto.setProductPrice(item.getPrice().doubleValue());
+                    productDto.setProductPrice(item.getPrice());
                     productDto.setProductStatus(ProductStatus.valueOf(item.getStatus()));
                     theProducts.add(productDto);
                 }
@@ -152,7 +152,7 @@ public class ProductRepository {
                         productDto.setProductID(String.valueOf(product.getId()));
                         productDto.setProductName(product.getProductName());
                         productDto.setProductImageURL("images/products/MeatlessPieces.png");
-                        productDto.setProductPrice(product.getPrice().doubleValue());
+                        productDto.setProductPrice(product.getPrice());
                         productDto.setProductStatus(ProductStatus.valueOf(product.getStatus()));
                         theProducts.add(productDto);
                     }
